backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.3

# Pydantic & Settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.27.0
aiofiles==23.2.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Image Processing & Computer Vision
pillow>=10.0.0
numpy>=1.24.0
roboflow>=1.2.0
replicate>=1.0.0

# Geospatial
shapely>=2.0.0
geojson>=3.1.0
pyproj>=3.6.0

# Fuzzy Matching
rapidfuzz>=3.5.0

# Web Scraping
beautifulsoup4>=4.12.0

# Object Storage
boto3>=1.34.0
